+++ resolwe/elastic/tests/test_app/migrations/0001_initial.py
@@ -27,6 +27,7 @@ class Migration(migrations.Migration):
                 ('name', models.CharField(max_length=30)),
                 ('field_process_type', models.CharField(max_length=100)),
                 ('number', models.IntegerField()),
+                ('date', models.DateTimeField(null=True)),
             ],
             options={
                 'permissions': (('view_testmodel', 'Can view model'), ('edit_testmodel', 'Can edit model')),