+++ docs/CHANGELOG.rst
@@ -24,6 +24,7 @@ Added
 - Improve log messages in executor and workload connectors
 - Add ``process_memory`` and ``process_cores`` fields to ``Data`` model and
   ``DataSerializer``
+- Support lookup expressions (``lt``, ``lte``, ``gt``, ``gte``) in ES viewsets
 
 Fixed
 -----